+++ app/decorators/property_decorator.rb
@@ -7,15 +7,15 @@ class PropertyDecorator < Draper::Decorator
       'house-flood-water'
     when 'franklin, trustee'
       'house-heart' # pro
-    else    
+    else
       lot_fees_paid? ? 'house-circle-check' : 'house-circle-xmark'
     end
   end
 
   def lot_numbers
     # Use leading zeros to create a "natural" sort
-    sorted_lots = lots.sort_by{|lot| format('%010s' % lot.lot_number) }
-    lots.collect{|lot| h.link_to(lot.lot_number, lot) }.join(', ').html_safe
+    sorted_lots = lots.decorate.sort_by{|lot| format('%010s' % lot.lot_number) }
+    sorted_lots.collect{|lot| h.link_to(lot.lot_number, lot) }.join(', ').html_safe
   end
 
   def residents_summary(type: :icons)