+++ app/models/resident.rb
@@ -6,7 +6,7 @@ class Resident < ApplicationRecord
   # List of searchable columns for this Model
   # ! this must be declared before pg_search_scope
   def self.searchable_columns
-    [:last_name]
+    [:last_name, :first_name]
   end
   pg_search_scope :search_by_all,
     # Reminder: first_name, email_address are encrypted