+++ db/migrate/20220724202314_add_abi_member_to_property.rb
@@ -1,0 +1,5 @@
+class AddAbiMemberToProperty < ActiveRecord::Migration[7.0]
+  def change
+    add_column :properties, :abi_member, :boolean
+  end
+end